Completed FirstLastVowel.java and updated MilesPerHour to accept 0 as hours

+++ Unit03/src/MilesPerHour.java
@@ -35,7 +35,7 @@ public class MilesPerHour
 
 	public void calcMPH()
 	{
-		mph = distance/(hours+(minutes/60));
+		mph = distance/(double)(hours+((double)minutes/60));
 	}
 
 	public void print()
